# packages/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: adres-pl — mikroserwis adresowy
  version: '2.0'
  description: |
    Podpowiedzi adresowe, walidacja i geokodowanie odwrotne na danych
    Państwowego Rejestru Granic (PRG) i TERYT.

    **Uwierzytelnianie: klucz API w nagłówku `X-API-Key`.** Klucz ma postać
    `adr_live_<32 znaki>_<6 znaków sumy>` i jest pokazywany **jeden raz**,
    przy wystawieniu — serwis przechowuje wyłącznie jego skrót.

    Limit zapytań liczony jest po **zweryfikowanym kliencie**, nie po kluczu
    ani po nagłówku: wszystkie klucze jednego klienta dzielą wspólny licznik.
    Ruch bez ważnego klucza podlega osobnemu, niższemu limitowi po adresie
    sieciowym. Limit obowiązuje **na instancję usługi** — przy wielu replikach
    łączna przepustowość jest odpowiednio wyższa.

    Sondy operacyjne (`/health`, `/ready`, `/metrics`, `/status`) pozostają
    bez uwierzytelniania: są potrzebne Kubernetesowi i Prometheuszowi, a odcina
    się je na warstwie sieciowej, nie w aplikacji.

    **Wersja danych** jest niezależna od wersji API. Zwracają ją `/v1/meta`
    oraz pola `wersjaDanych` w odpowiedziach wyszukiwania i walidacji.
  license:
    name: Dane PRG bez licencji (art. 40c ust. 5 Prawa geodezyjnego)
servers:
  - url: http://localhost:3000
    description: Instalacja lokalna

# Obowiazuje wszystkie operacje POZA tymi, ktore maja jawne `security: []`
# (sondy operacyjne). Wpisanie tego globalnie, a nie przy kazdej operacji,
...
# napisac `security: []`, zeby go otworzyc.
security:
  - ApiKey: []
...
paths:
...
  /v1/locality/{simc}:
    get:
      tags: [search]
      summary: Szczegóły miejscowości
      description: |
        Pole `ma_ulice` steruje formularzem: gdy `false`, pole ulicy należy
        **ukryć**, a nie zablokować — numer odnosi się wtedy bezpośrednio
        do miejscowości.
      parameters:
        - { name: simc, in: path, required: true, schema: { type: string, pattern: '^[0-9]{7}$' } }
      responses:
        '200':
          description: Miejscowość wraz z hierarchią administracyjną.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Miejscowosc' }
        '404': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimitExceeded' }
...
components:
  securitySchemes:
    OperatorToken:
      type: http
      scheme: bearer
      description: |
        Token operatora do zarządzania klientami i kluczami. **Odrębny od kluczy
        klienckich**: klucz `adr_live_*` nie otwiera żadnej trasy `/admin` —
        inaczej dowolny klient wystawiłby sobie klucz bez limitów.
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Klucz klienta. **Wyłącznie w nagłówku** — klucz podany w query stringu
        jest traktowany jak jego brak, bo trafiłby do access logu ingressu,
        do nagłówka `Referer` i do historii przeglądarki.
  responses:
...
    Unauthenticated:
      description: |
        Brak klucza albo klucz nieprawidłowy.

        Zły format, zła suma kontrolna i klucz nieznany dają **identyczną**
        odpowiedź — rozróżnienie zamieniłoby ją w wyrocznię dla zgadującego.
      content:
        application/json:
          schema:
            type: object
            required: [error, code]
            properties:
              error: { type: string, example: 'Klucz API nieprawidlowy.' }
              code: { type: string, enum: [BRAK_KLUCZA, NIEPRAWIDLOWY] }
    Forbidden:
      description: |
        Klucz rozpoznany, ale nieczynny. Kod mówi wprost, co się stało —
        na tym etapie klient udowodnił już posiadanie całego sekretu, więc
        nie ma czego chronić, a integrator musi wiedzieć, co zrobić.
      content:
        application/json:
          schema:
            type: object
            required: [error, code]
            properties:
              error: { type: string, example: 'Klucz API wygasl.' }
              code: { type: string, enum: [NIEWAZNY_JESZCZE, WYGASLY, UNIEWAZNIONY, ZAWIESZONY] }
...
    BadRequest:
      description: Żądanie poprawne składniowo, ale niewykonalne.
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string, example: 'Podaj ulicId albo simc.' }
    RateLimitExceeded:
      description: |
        Przekroczony limit zapytań. Limit liczony jest po adresie sieciowym
        klienta, domyślnie 600 żądań na minutę.
  schemas:
...
    Miejscowosc:
      type: object
      properties:
        simc: { type: string }
        name: { type: string }
        kind: { type: integer }
        rodzaj_nazwa: { type: [string, 'null'], example: miasto }
        ma_ulice: { type: boolean }
        liczba_punktow: { type: integer }
        terc_gminy: { type: string }
        gmina: { type: string }
        powiat: { type: [string, 'null'] }
        voivodeship: { type: [string, 'null'] }
        lat: { type: [number, 'null'] }
        lon: { type: [number, 'null'] }
...